requirements.txt
aiohttp==3.13.3
python-socketio==5.16.0
